4.21 and fixing problem of blog creation where user not shows

+++ osa4/bloglist/controllers/blogs.js
@@ -18,8 +18,6 @@ blogsRouter.get('/', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   try {
-    const blog = new Blog(request.body)
-
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token invalid' })
@@ -28,6 +26,11 @@ blogsRouter.post('/', async (request, response, next) => {
     if (!user) {
       return response.status(400).json({ error: 'UserId missing or not valid' })
     }
+
+    const blog = new Blog({
+      ...request.body,
+      user: user._id
+    })
 
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
@@ -41,6 +44,21 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: 'blog missing' })
+    }
+    console.log('blog.user.toString():', blog.user.toString())
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(401).json({ error: 'adder of the blog is not the same as the user' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (error) {
